+++ R/rxnorm_map.R
@@ -7,6 +7,9 @@
 #' that will be used to generate the SQL queries to derive RxNorm maps.
 #' @param url Source URL. It is not hardcoded in case the URL changes in the future.
 #'  Default: 'https://lhncbc.nlm.nih.gov/RxNav/applications/RxNavViews.html#label:appendix'
+#' @details
+#' The tibble parsed from the URL is cached indefinitely. This is to avoid losing
+#' this dataframe in the event that the source documentation is ever lost.
 #' @return
 #' Tibble of a machine-readable version of what is read at the `url` parameter.
 #' @rdname get_rxnorm_paths
@@ -16,10 +19,30 @@
 #' @import tidyr
 #' @import dplyr
 #' @import pg13
+#' @import tibble
+#' @import purrr
+#' @import R.cache
+#' @import secretary
 #' @importFrom rlang parse_expr
 #' @importFrom glue glue
 get_rxnorm_paths <-
         function(url = "https://lhncbc.nlm.nih.gov/RxNav/applications/RxNavViews.html#label:appendix") {
+
+          cache_file <-
+          R.cache::findCache(key = list(url = url),
+                             dirs = "metathesaurus")
+
+
+          if (!is.null(cache_file)) {
+
+            cached_datetime <- file.info(cache_file)$ctime
+            secretary::typewrite(glue::glue("Loading cached results from {prettyunits::time_ago(cached_datetime)}."))
+            R.cache::loadCache(key = list(url = url),
+                               dirs = "metathesaurus")
+
+          } else {
+
+
                 input <-
                         read_html(url) %>%
                         rvest::html_table() %>%
@@ -96,20 +119,134 @@ get_rxnorm_paths <-
                                 into = c("start", "end"),
                                 regex = "([A-Z]{2,}) => ([A-Z]{2,})")
 
+                R.cache::saveCache(object = output_list2,
+                                   key = list(url = url),
+                                   dirs = "metathesaurus")
+
                 output_list2
+          }
+
         }
 
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
-#' @param conn PARAM_DESCRIPTION
-#' @param conn_fun PARAM_DESCRIPTION, Default: 'pg13::local_connect()'
-#' @param start_arg PARAM_DESCRIPTION
-#' @param end_arg PARAM_DESCRIPTION
-#' @param verbose PARAM_DESCRIPTION, Default: TRUE
-#' @param render_sql PARAM_DESCRIPTION, Default: TRUE
-#' @param render_only PARAM_DESCRIPTION, Default: FALSE
-#' @param checks PARAM_DESCRIPTION, Default: ''
-#' @return OUTPUT_DESCRIPTION
+
+
+#' @title
+#' Write RxNorm Path Lookup
+#'
+#' @description
+#' Write the tibble returned by `get_rxnorm_paths()` to a
+#' Postgres table.
+#'
+#' @inheritParams pkg_args
+#' @inheritParams get_rxnorm_paths
+#' @inheritParams pg13::write_table
+#' @rdname write_rxnorm_path_lookup
+#' @family RxNorm Map
+#' @export
+#' @importFrom rlang parse_expr
+#' @importFrom pg13 dc write_table
+
+write_rxnorm_path_lookup <-
+  function(conn,
+           conn_fun = "pg13::local_connect()",
+           schema = "rxclass",
+           table_name = "lookup_rxnorm_paths",
+           url = "https://lhncbc.nlm.nih.gov/RxNav/applications/RxNavViews.html#label:appendix",
+           verbose = TRUE,
+           render_sql = TRUE,
+           render_only = FALSE,
+           checks = "") {
+
+
+
+    out <-
+      get_rxnorm_paths(url = url) %>%
+      dplyr::group_by(start, end) %>%
+      dplyr::arrange(as.integer(path_level),
+                     .by_group = TRUE) %>%
+      tibble::rowid_to_column()
+
+
+    if (missing(conn)) {
+      conn <- eval(rlang::parse_expr(conn_fun))
+      on.exit(expr = pg13::dc(conn = conn), add = TRUE,
+              after = TRUE)
+    }
+
+
+    tmp_csv <- tempfile(fileext = ".csv")
+
+    readr::write_csv(x = out,
+                     file = tmp_csv,
+                     na = "",
+                     col_names = TRUE)
+
+    on.exit(expr = unlink(tmp_csv),
+            add = TRUE,
+            after = TRUE)
+
+    pg13::send(
+      conn = conn,
+      sql_statement =
+        glue::glue(
+        "
+        DROP TABLE IF EXISTS {schema}.tmp_{table_name};
+        CREATE TABLE {schema}.tmp_{table_name} (
+          rowid int NOT NULL,
+          rxnorm_start_tty varchar(10) NOT NULL,
+          rxnorm_end_tty varchar(10) NOT NULL,
+          path_level int NOT NULL,
+          from_tty varchar(10) NOT NULL,
+          to_tty varchar(10) NOT NULL
+        );
+
+        COPY {schema}.tmp_{table_name}
+        FROM '{tmp_csv}'
+        CSV HEADER QUOTE E'\\b';
+
+        DROP TABLE IF EXISTS {schema}.{table_name};
+        CREATE TABLE {schema}.{table_name} AS (
+          SELECT
+            tmp.rxnorm_start_tty,
+            tmp.rxnorm_end_tty,
+            tmp.path_level,
+            tmp.from_tty,
+            tmp.to_tty
+          FROM {schema}.tmp_{table_name} tmp
+          ORDER BY tmp.rowid
+        );
+
+        DROP TABLE {schema}.tmp_{table_name};
+        ",
+        checks = checks,
+        verbose = verbose,
+        render_sql = render_sql,
+        render_only = render_only)
+    )
+
+  }
+
+
+
+
+#' @title
+#' Get the RxNorm Map Between 2 TTY
+#'
+#' @description
+#' Get the RxNorm Map Between 2 TTY. This map
+#' ignores any variability from the rel/rela.
+#' @inheritParams pkg_args
+#' @inheritParams pg13::query
+#' @param from_tty The RxNorm TTY to start from.
+#' @param to_tty   The RxNorm TTY to end at.
+#' @param full_path By default, regardless of path length, only the `from_tty` and `to_tty`
+#' values are returned. If TRUE, the entire path is returned instead.
+#' @return
+#' Tibble of the mappings between the `aui`, `code`, and `str` between
+#' the two start and end tty arguments. If `full_path` is TRUE, the intermediate tty
+#' values are also provided. Every set is prefixed with the tty value. The returned map
+#' is derived from a full join and therefore, blank mappings between tty are also included
+#' in the tibble.
 #' @details DETAILS
 #' @examples
 #' \dontrun{
@@ -123,6 +260,7 @@ get_rxnorm_paths <-
 #'  \code{\link[cli]{cli_abort}}
 #'  \code{\link[pg13]{query}}
 #' @rdname get_rxnorm_map
+#' @family RxNorm Map
 #' @export
 #' @importFrom dplyr filter
 #' @importFrom glue glue
@@ -131,30 +269,57 @@ get_rxnorm_paths <-
 get_rxnorm_map <-
         function(conn,
                  conn_fun = "pg13::local_connect()",
-                 start_arg,
-                 end_arg,
+                 from_tty = c('BN', 'BPCK', 'DF', 'DFG', 'ET', 'GPCK', 'IN', 'MIN', 'PIN', 'PSN', 'SBD', 'SBDC', 'SBDF', 'SBDG', 'SCD', 'SCDC', 'SCDF', 'SCDG', 'SY', 'TMSY'),
+                 to_tty = c('BN', 'BPCK', 'DF', 'DFG', 'ET', 'GPCK', 'IN', 'MIN', 'PIN', 'PSN', 'SBD', 'SBDC', 'SBDF', 'SBDG', 'SCD', 'SCDC', 'SCDF', 'SCDG', 'SY', 'TMSY'),
+                 full_path = FALSE,
+                 schema = "mth",
                  verbose = TRUE,
                  render_sql = TRUE,
                  render_only = FALSE,
                  checks = "") {
 
+            # Match Arguments
+            from_tty <-
+            match.arg(arg = from_tty,
+                      choices = c('BN', 'BPCK', 'DF', 'DFG', 'ET', 'GPCK', 'IN', 'MIN', 'PIN', 'PSN', 'SBD', 'SBDC', 'SBDF', 'SBDG', 'SCD', 'SCDC', 'SCDF', 'SCDG', 'SY', 'TMSY'),
+                      several.ok = FALSE)
+
+            to_tty <-
+              match.arg(arg = to_tty,
+                        choices = c('BN', 'BPCK', 'DF', 'DFG', 'ET', 'GPCK', 'IN', 'MIN', 'PIN', 'PSN', 'SBD', 'SBDC', 'SBDF', 'SBDG', 'SCD', 'SCDC', 'SCDF', 'SCDG', 'SY', 'TMSY'),
+                        several.ok = FALSE)
+
+            if (identical(from_tty, to_tty)) {
+
+              cli::cli_alert_danger("`from_tty` and `to_tty` are the same value. '{from_tty}' cannot be mapped to itself.")
+
+
+
+            } else {
+
+              if (missing(conn)) {
+                conn <- eval(rlang::parse_expr(conn_fun))
+                on.exit(expr = pg13::dc(conn = conn), add = TRUE,
+                        after = TRUE)
+              }
+
                 rxnorm_concept_map <-
-                        get_rxnorm_concept_map()
+                        get_rxnorm_paths()
                 Sys.sleep(1)
                 run_map <-
                         rxnorm_concept_map %>%
-                        dplyr::filter(start %in% start_arg,
-                                      end %in% end_arg) %>%
+                        dplyr::filter(start %in% from_tty,
+                                      end %in% to_tty) %>%
                         distinct() %>%
                         arrange(as.integer(path_level))
 
                 if (nrow(run_map)==0) {
-                        stop(glue::glue("A path starting from '{start_arg}' and ending at '{end_arg}' does not exist."))
+                        stop(glue::glue("A path starting from '{from_tty}' and ending at '{to_tty}' does not exist."))
                 }
 
                 max_path_level <-
                         max(run_map$path_level)
-                cli::cli_inform("Path from '{start_arg}' to '{end_arg}' is {max_path_level} level{?s} in length.")
+                cli::cli_inform("Path from '{from_tty}' to '{to_tty}' is {max_path_level} level{?s} in length.")
 
                 output <- list()
                 for (i in 1:nrow(run_map)) {
@@ -171,10 +336,10 @@ get_rxnorm_map <-
      m2.aui AS {to}_aui,
      m2.code AS {to}_code,
      m2.str  AS {to}_str
-    FROM umls.MRCONSO m1
-    INNER JOIN umls.MRREL r
+    FROM {schema}.MRCONSO m1
+    INNER JOIN {schema}.MRREL r
     ON r.aui1 = m1.aui
-    INNER JOIN umls.MRCONSO m2
+    INNER JOIN {schema}.MRCONSO m2
     ON r.aui2 = m2.aui
     WHERE
       m1.sab = 'RXNORM' AND
@@ -196,37 +361,59 @@ get_rxnorm_map <-
                                             checks = checks)
 
                 }
-                final_output <-
-                        output %>%
-                        map(function(x) x$value)
+                final_output <- output
 
+                if (full_path) {
+
+                  final_output %>%
+                    reduce(full_join) %>%
+                    distinct()
+
+
+                } else {
 
                 final_output %>%
                         reduce(full_join) %>%
-                        select(starts_with(tolower(start_arg)),
-                               starts_with(tolower(end_arg))) %>%
+                        select(starts_with(tolower(from_tty)),
+                               starts_with(tolower(to_tty))) %>%
                         distinct()
+
+                }
+            }
         }
 
 
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
-
-#' @return OUTPUT_DESCRIPTION
-#' @details DETAILS
-#' @examples
-#' \dontrun{
-#' if(interactive()){
-#'  #EXAMPLE1
-#'  }
-#' }
-#' @seealso
-#'  \code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter_all}}
+#' @title
+#' Get RxNorm Ingredient Map
+#'
+#' @description
+#' Iterate on all the possible paths
+#' out of the RxNorm Ingredient concepts
+#' using the `get_rxnorm_map` function.
+#'
+#' @return
+#' Tibble of each unique combination of
+#' from tty 'IN' and an ending tty with an extra field that
+#' provides the given tty value. The final tibble
+#' as includes 'IN' mappings to itself.
+#'
+#' @inheritParams  get_rxnorm_map
 #' @rdname get_rxnorm_ingredient_map
+#' @family RxNorm Map
 #' @export
-#' @importFrom dplyr filter transmute filter_at
+#' @import dplyr
+#' @import tidyr
+#' @import purrr
+#' @import pg13
+
 get_rxnorm_ingredient_map <-
-        function() {
+        function(conn,
+                 conn_fun = "pg13::local_connect()",
+                 schema = "mth",
+                 verbose = TRUE,
+                 render_sql = TRUE,
+                 render_only = FALSE,
+                 checks = "") {
                 rxnorm_concept_map <-
                         get_rxnorm_paths()
 
@@ -234,22 +421,38 @@ get_rxnorm_ingredient_map <-
                         rxnorm_concept_map %>%
                         dplyr::filter(start == 'IN')
 
-                end_args <-
+                to_ttys <-
                         unique(ingr_rxnorm_concept_map$end)
+
+                if (missing(conn)) {
+                  conn <- eval(rlang::parse_expr(conn_fun))
+                  on.exit(expr = pg13::dc(conn = conn), add = TRUE,
+                          after = TRUE)
+                }
+
 
                 rxnorm_ingredient_map <-
                         vector(mode = "list",
-                               length = length(end_args))
+                               length = length(to_ttys))
                 names(rxnorm_ingredient_map) <-
-                        end_args
+                        to_ttys
 
-                for (end_arg in end_args) {
+                for (to_tty in to_ttys) {
 
-                        rxnorm_ingredient_map[[end_arg]] <-
-                                get_rxnorm_map(start_arg =  "IN",
-                                               end_arg = end_arg) %>%
+                        rxnorm_ingredient_map[[to_tty]] <-
+                                get_rxnorm_map(from_tty =  "IN",
+                                               to_tty = to_tty,
+                                               conn = conn,
+                                               start_arg = start_arg,
+                                               end_arg = end_arg,
+                                               full_path = full_path,
+                                               schema = schema,
+                                               verbose = verbose,
+                                               render_sql = render_sql,
+                                               render_only = render_only,
+                                               checks = checks) %>%
                                 rename_all(str_remove_all,
-                                           pattern = paste0(tolower(end_arg), "_"))
+                                           pattern = paste0(tolower(to_tty), "_"))
                 }
 
                 rxnorm_ingredient_map2 <-
@@ -280,5 +483,8 @@ get_rxnorm_ingredient_map <-
                         dplyr::filter_at(vars(c(aui, code, str)),
                                          all_vars(!is.na(.))) %>%
                         distinct()
+
+
+                final_rxnorm_ingredient_map
 
         }